base: New method for IOC build
This implementation was introduced due to the need to build IOCs
using only internal modules from epics-in-docker. With this new
approach, building this type of IOC requires only the definition of
the required libraries and database files in the YAML configuration
file.

As a consequence, IOCs that are currently built together with the
modules will be removed in the future.

During development, it was identified that the direct use of the
DBD, LIBS, and DB variables was not feasible, as it caused variable
conflicts during the make execution.
Initially, the use of the "sed" command was considered to replace
and inject parameters into the Makefile; however, after further
discussion, this approach was deemed unsustainable in the long term.
Therefore, a static Makefile approach was adopted, simplifying
maintenance.

The creation of a new Dockerfile was also evaluated. However, after
analyzing the root Dockerfile, it was observed that adding a single
command line was sufficient to address the issue, avoiding code
duplication and preserving compatibility with the static-link and
dynamic-link targets.